WIP: sync code generation macros
This is a pre-removal of sync code.

Turns out that it doesn't make much sense to internally control RwLocks
and Mutex because in some cases users might require different types of
access with different characteristics, so I believe a better approach
would be making the core impl. as quick and simple as possible and then
users of the library can wrap it in RwLock or Mutex as they see fit.

Otherwise, implementation is super complicated and difficult to modify
for those who aren't acquainted with the codebase. Given the permissive
nature of the license, and how niche the synchronised use case is, I
think it's a better choice to provide a simpler implementation which can
then be re-vendored and modified to fit specific requirements if
wrapping the structure in RwLock/Mutex isn't good enough.

I'm pushing this for future reference, but most of the changes in this
commit will be reverted.

Signed-off-by: Tin Svagelj <tin.svagelj@live.com>